Put each sentence on a new line
It is easier for git to figure out what changes have been made if
each sentence is put on a new line. The linebreaks do not make any
difference to markdown, so they are still rendered the same, so
git will check diffs line by line, _ie_ sentence by sentence.

One could  of course put each _word_ on a new line, but then you're
going to have a hard time reading the file !
This is a bit of a best-practice tradeoff